+++ Config.php
@@ -16,7 +16,7 @@ return [
 	
 	'default_password' => 'P@ssword',
 	
-	'storage_path' => __DIR__ . '/storage',
+	'storage_path' => __DIR__ . '/Storage',
 	
 	'timezone' => 'UTC',
 	
@@ -25,6 +25,10 @@ return [
 	'refresh_expiration' => '+30 days',
 	
 	'refresh_absolute_expiration' => '+90 days',
+	
+	'calendar_prodid' => '-//Asprovalta Seaside Home Calendar//EN',
+	
+	'calendar_name' => 'Asprovalta Seaside Home Availability',
 	
 	'encryption_key' => hex2bin('75150a7ac135b7cf922c51a16c6d7f150ec048c8f4ae45e7558543b5a31a2e69')
 ];